chore(core): split 10 oversized TypeScript files to meet 500-line standard (SMI-2741)
Split 10 files that exceeded the 500-line engineering standard into adjacent
companion files. Each original file re-exports everything it previously exported
so there are no public API changes.

Files split and their companions:
- context.ts (705→333) → context.types.ts, context.helpers.ts, context.async.ts
- errorFormatter.ts (547→212) → errorFormatter.types.ts, errorFormatter.builders.ts
- QuarantineService.ts (554→327) → QuarantineService.multiapproval.ts
- TaskRunner.ts (542→429) → TaskRunner.types.ts, TaskRunner.process.ts
- client.ts (542→490) → client.health.ts
- MultiLLMProvider.ts (589→471) → MultiLLMProvider.metrics.ts, MultiLLMProvider.selection.ts
- recommend.ts (546→431) → recommend.format.ts
- SessionManager.ts (515→294) → SessionManager.memory.ts
- install.ts (560→492) → install.tool.ts, install.optimize.ts
- PatternStore.ts (519→484) → PatternStore.helpers.ts (extended)

All quality gates pass: build, typecheck, lint (0 warnings), tests (5680 passed),
audit:standards (100%, all files under 500 lines).
