+++ packages/core/src/session/SessionManager.ts
@@ -2,6 +2,7 @@
  * Session Manager for Claude-Flow Memory Integration
  * SMI-641: Session ID Storage in Claude-Flow Memory
  * SMI-1518: V3 API Migration - Use direct API calls instead of spawn
+ * SMI-2741: Split to meet 500-line standard
  *
  * Manages session lifecycle with persistent storage in claude-flow memory
  * to enable context restoration across sessions.
@@ -15,19 +16,27 @@ import type { CommandExecutor, MemoryResult, SessionOptions } from './SessionMan
 import { sanitizeSessionData } from './SessionManager.types.js'
 
 // Import helpers
+import { MEMORY_KEYS, DefaultCommandExecutor } from './SessionManager.helpers.js'
+
+// SMI-2741: Memory and hook operations extracted to companion file
 import {
-  getClaudeFlowMemory,
-  getClaudeFlowMcp,
-  MEMORY_KEYS,
-  USE_V3_API,
-  MEMORY_NAMESPACE,
-  validateMemoryKey,
-  DefaultCommandExecutor,
-} from './SessionManager.helpers.js'
+  storeMemoryEntry,
+  retrieveMemoryEntry,
+  deleteMemoryEntry,
+  runPreTaskHook,
+  runPostTaskHook,
+} from './SessionManager.memory.js'
 
 // Re-export only public API types (SMI-1718: trimmed internal exports)
 export type { CommandExecutor, MemoryResult, SessionOptions } from './SessionManager.types.js'
 export { DefaultCommandExecutor } from './SessionManager.helpers.js'
+export {
+  storeMemoryEntry,
+  retrieveMemoryEntry,
+  deleteMemoryEntry,
+  runPreTaskHook,
+  runPostTaskHook,
+} from './SessionManager.memory.js'
 
 /**
  * Session Manager for claude-flow memory integration
@@ -109,7 +118,7 @@ export class SessionManager {
 
     // Run pre-task hook if description provided
     if (options.description) {
-      await this.runPreTaskHook(options.description)
+      await runPreTaskHook(options.description, this.executor)
     }
 
     this.currentSession = session
@@ -136,9 +145,10 @@ export class SessionManager {
       }
 
       // SMI-676: Store checkpoint data FIRST (before updating session)
-      const checkpointResult = await this.storeMemory(
+      const checkpointResult = await storeMemoryEntry(
         checkpoint.memoryKey,
-        JSON.stringify(checkpoint)
+        JSON.stringify(checkpoint),
+        this.executor
       )
 
       if (!checkpointResult.success) {
@@ -165,7 +175,7 @@ export class SessionManager {
         this.currentSession.lastActivity = previousLastActivity
 
         // Clean up the checkpoint from memory
-        await this.deleteMemory(checkpoint.memoryKey)
+        await deleteMemoryEntry(checkpoint.memoryKey, this.executor)
 
         throw err
       }
@@ -205,7 +215,7 @@ export class SessionManager {
     await this.storeSession(this.currentSession)
 
     // Run post-task hook
-    await this.runPostTaskHook(this.currentSession.sessionId)
+    await runPostTaskHook(this.currentSession.sessionId, this.executor)
 
     // Clear current session pointer
     await this.clearCurrentSession()
@@ -225,7 +235,7 @@ export class SessionManager {
    */
   async getSession(sessionId: string): Promise<SessionData | null> {
     const memoryKey = `${MEMORY_KEYS.SESSION_PREFIX}${sessionId}`
-    const result = await this.retrieveMemory(memoryKey)
+    const result = await retrieveMemoryEntry(memoryKey, this.executor)
 
     if (!result.success || !result.data) {
       return null
@@ -242,7 +252,7 @@ export class SessionManager {
    * Get the ID of the current session from memory
    */
   async getCurrentSessionId(): Promise<string | null> {
-    const result = await this.retrieveMemory(MEMORY_KEYS.CURRENT)
+    const result = await retrieveMemoryEntry(MEMORY_KEYS.CURRENT, this.executor)
 
     if (!result.success || !result.data) {
       return null
@@ -262,254 +272,20 @@ export class SessionManager {
   private async storeSession(session: SessionData): Promise<MemoryResult> {
     const memoryKey = `${MEMORY_KEYS.SESSION_PREFIX}${session.sessionId}`
     const sanitized = sanitizeSessionData(session)
-    return this.storeMemory(memoryKey, JSON.stringify(sanitized))
+    return storeMemoryEntry(memoryKey, JSON.stringify(sanitized), this.executor)
   }
 
   /**
    * Set the current session pointer
    */
   private async setCurrentSession(sessionId: string): Promise<MemoryResult> {
-    return this.storeMemory(MEMORY_KEYS.CURRENT, JSON.stringify({ sessionId }))
+    return storeMemoryEntry(MEMORY_KEYS.CURRENT, JSON.stringify({ sessionId }), this.executor)
   }
 
   /**
    * Clear the current session pointer
    */
   private async clearCurrentSession(): Promise<MemoryResult> {
-    return this.deleteMemory(MEMORY_KEYS.CURRENT)
-  }
-
-  /**
-   * Store data in claude-flow memory
-   *
-   * SMI-674: Uses spawn() with argument array to prevent command injection
-   * SMI-1518: V3 API Migration - Use direct storeEntry() when available
-   */
-  private async storeMemory(key: string, value: string): Promise<MemoryResult> {
-    if (!validateMemoryKey(key)) {
-      return { success: false, error: 'Invalid memory key' }
-    }
-
-    // SMI-1518, SMI-1609: Try V3 direct API first if enabled
-    if (USE_V3_API) {
-      try {
-        const memoryModule = await getClaudeFlowMemory()
-        if (memoryModule?.storeEntry) {
-          const result = await memoryModule.storeEntry({
-            key,
-            value,
-            namespace: MEMORY_NAMESPACE,
-          })
-          if (result.success) {
-            return { success: true }
-          }
-          console.warn(`V3 storeEntry failed: ${result.error}, falling back to spawn`)
-        }
-      } catch (error) {
-        console.warn(`V3 storeEntry exception: ${error}, falling back to spawn`)
-      }
-    }
-
-    try {
-      const args = ['claude-flow', 'memory', 'store', '--key', key, '--value', value]
-
-      if (this.executor.spawn) {
-        await this.executor.spawn('npx', args)
-      } else {
-        const escapedValue = value.replace(/'/g, "'\\''")
-        const command = `npx claude-flow memory store --key "${key}" --value '${escapedValue}'`
-        await this.executor.execute(command)
-      }
-      return { success: true }
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : String(error),
-      }
-    }
-  }
-
-  /**
-   * Retrieve data from claude-flow memory
-   *
-   * SMI-674: Uses spawn() with argument array to prevent command injection
-   * SMI-1518: V3 API Migration - Use direct getEntry() when available
-   */
-  private async retrieveMemory(key: string): Promise<MemoryResult> {
-    if (!validateMemoryKey(key)) {
-      return { success: false, error: 'Invalid memory key' }
-    }
-
-    // SMI-1518, SMI-1609: Try V3 direct API first if enabled
-    if (USE_V3_API) {
-      try {
-        const memoryModule = await getClaudeFlowMemory()
-        if (memoryModule?.getEntry) {
-          const result = await memoryModule.getEntry({
-            key,
-            namespace: MEMORY_NAMESPACE,
-          })
-          if (result.success && result.found && result.entry) {
-            return { success: true, data: result.entry.content }
-          }
-          if (result.success && !result.found) {
-            return { success: false, error: 'Key not found' }
-          }
-          console.warn(`V3 getEntry failed: ${result.error}, falling back to spawn`)
-        }
-      } catch (error) {
-        console.warn(`V3 getEntry exception: ${error}, falling back to spawn`)
-      }
-    }
-
-    try {
-      const args = ['claude-flow', 'memory', 'get', '--key', key]
-
-      let stdout: string
-      if (this.executor.spawn) {
-        const result = await this.executor.spawn('npx', args)
-        stdout = result.stdout
-      } else {
-        const command = `npx claude-flow memory get --key "${key}"`
-        const result = await this.executor.execute(command)
-        stdout = result.stdout
-      }
-      return { success: true, data: stdout.trim() }
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : String(error),
-      }
-    }
-  }
-
-  /**
-   * Delete data from claude-flow memory
-   *
-   * SMI-674: Uses spawn() with argument array to prevent command injection
-   * SMI-1518: V3 API Migration - Use callMCPTool('memory/delete') when available
-   */
-  private async deleteMemory(key: string): Promise<MemoryResult> {
-    if (!validateMemoryKey(key)) {
-      return { success: false, error: 'Invalid memory key' }
-    }
-
-    // SMI-1518, SMI-1609: Try V3 MCP API first if enabled
-    if (USE_V3_API) {
-      try {
-        const mcpModule = await getClaudeFlowMcp()
-        if (mcpModule?.callMCPTool) {
-          const result = (await mcpModule.callMCPTool('memory/delete', { key })) as {
-            success: boolean
-            deleted: boolean
-          }
-          if (result.success) {
-            return { success: true }
-          }
-          console.warn(`V3 memory/delete failed, falling back to spawn`)
-        }
-      } catch (error) {
-        const mcpModule = await getClaudeFlowMcp()
-        const MCPClientError = mcpModule?.MCPClientError
-        if (!MCPClientError || !(error instanceof MCPClientError)) {
-          console.warn(`V3 memory/delete exception: ${error}, falling back to spawn`)
-        }
-      }
-    }
-
-    try {
-      const args = ['claude-flow', 'memory', 'delete', '--key', key]
-
-      if (this.executor.spawn) {
-        await this.executor.spawn('npx', args)
-      } else {
-        const command = `npx claude-flow memory delete --key "${key}"`
-        await this.executor.execute(command)
-      }
-      return { success: true }
-    } catch {
-      // Ignore delete errors (key may not exist)
-      return { success: true }
-    }
-  }
-
-  /**
-   * Run pre-task hook
-   * SMI-674: Uses spawn() with argument array to prevent command injection
-   * SMI-1518: V3 API Migration - Use callMCPTool('hooks/pre-task') when available
-   */
-  private async runPreTaskHook(description: string): Promise<void> {
-    // SMI-1518, SMI-1609: Try V3 MCP API first if enabled
-    if (USE_V3_API) {
-      try {
-        const mcpModule = await getClaudeFlowMcp()
-        if (mcpModule?.callMCPTool) {
-          await mcpModule.callMCPTool('hooks/pre-task', {
-            description,
-            memoryKey: 'session/current',
-          })
-          return
-        }
-      } catch {
-        // V3 API not available or failed, fall back to spawn
-      }
-    }
-
-    try {
-      const args = [
-        'claude-flow',
-        'hooks',
-        'pre-task',
-        '--description',
-        description,
-        '--memory-key',
-        'session/current',
-      ]
-
-      if (this.executor.spawn) {
-        await this.executor.spawn('npx', args)
-      } else {
-        const escapedDesc = description.replace(/'/g, "'\\''")
-        const command = `npx claude-flow hooks pre-task --description '${escapedDesc}' --memory-key "session/current"`
-        await this.executor.execute(command)
-      }
-    } catch {
-      // Hooks are optional, don't fail if they don't work
-    }
-  }
-
-  /**
-   * Run post-task hook
-   * SMI-674: Uses spawn() with argument array to prevent command injection
-   * SMI-1518: V3 API Migration - Use callMCPTool('hooks/post-task') when available
-   */
-  private async runPostTaskHook(taskId: string): Promise<void> {
-    // SMI-1518, SMI-1609: Try V3 MCP API first if enabled
-    if (USE_V3_API) {
-      try {
-        const mcpModule = await getClaudeFlowMcp()
-        if (mcpModule?.callMCPTool) {
-          await mcpModule.callMCPTool('hooks/post-task', {
-            taskId,
-          })
-          return
-        }
-      } catch {
-        // V3 API not available or failed, fall back to spawn
-      }
-    }
-
-    try {
-      const args = ['claude-flow', 'hooks', 'post-task', '--task-id', taskId]
-
-      if (this.executor.spawn) {
-        await this.executor.spawn('npx', args)
-      } else {
-        const command = `npx claude-flow hooks post-task --task-id "${taskId}"`
-        await this.executor.execute(command)
-      }
-    } catch {
-      // Hooks are optional, don't fail if they don't work
-    }
+    return deleteMemoryEntry(MEMORY_KEYS.CURRENT, this.executor)
   }
 }
