+++ packages/mcp-server/src/tools/recommend.ts
@@ -5,12 +5,13 @@
  * @see SMI-602: Integrate semantic matching with EmbeddingService
  * @see SMI-604: Add trigger phrase overlap detection
  * @see SMI-1837: Include local skills in recommendations (parallel search)
+ * @see SMI-2741: Split to meet 500-line standard
  */
 
 import { SkillMatcher, OverlapDetector, trackEvent } from '@skillsmith/core'
 import type { ToolContext } from '../context.js'
 import { getInstalledSkills } from '../utils/installed-skills.js'
-import { mapTrustTierFromDb, getTrustBadge } from '../utils/validation.js'
+import { mapTrustTierFromDb } from '../utils/validation.js'
 
 // Import types
 import {
@@ -28,6 +29,9 @@ import {
   isSkillCollection,
 } from './recommend.helpers.js'
 
+// SMI-2741: Formatting and deduplication extracted to companion file
+import { mergeAndDeduplicateRecommendations } from './recommend.format.js'
+
 // SMI-1837: Import local skill search for parallel querying
 import { getLocalIndexer } from './LocalSkillSearch.js'
 import type { LocalSkill } from '../indexer/LocalIndexer.js'
@@ -40,6 +44,9 @@ export {
   type RecommendResponse,
   type SkillRecommendation,
 } from './recommend.types.js'
+
+// Re-export formatting utilities (SMI-2741)
+export { formatRecommendations, mergeAndDeduplicateRecommendations } from './recommend.format.js'
 
 /**
  * SMI-1837: Convert a LocalSkill to SkillRecommendation format
@@ -95,66 +102,6 @@ async function searchLocalSkillsForRecommend(
     console.warn('[skillsmith] Local skill search for recommend failed:', (error as Error).message)
     return []
   }
-}
-
-/**
- * SMI-1837: Merge and deduplicate API and local skill recommendations
- * API results take priority over local results with the same name
- * @param apiResults - Results from API
- * @param localResults - Results from local skill search
- * @param limit - Maximum combined results
- * @returns Merged and deduplicated recommendations
- */
-function mergeAndDeduplicateRecommendations(
-  apiResults: SkillRecommendation[],
-  localResults: SkillRecommendation[],
-  limit: number
-): SkillRecommendation[] {
-  // Build a Set of names from API results for deduplication
-  const apiSkillNames = new Set(apiResults.map((r) => r.name.toLowerCase()))
-
-  // Also track skill IDs (without the author prefix)
-  const apiSkillIdNames = new Set(
-    apiResults.map((r) => r.skill_id.split('/').pop()?.toLowerCase() || '')
-  )
-
-  // Filter local results to exclude duplicates
-  const uniqueLocalResults = localResults.filter((local) => {
-    const localName = local.name.toLowerCase()
-    const localIdName = local.skill_id.split('/').pop()?.toLowerCase() || ''
-
-    // Exclude if name matches an API result
-    if (apiSkillNames.has(localName)) {
-      return false
-    }
-
-    // Exclude if ID name matches an API result
-    if (apiSkillIdNames.has(localIdName)) {
-      return false
-    }
-
-    // Exclude if local name is contained in or contains an API skill name
-    for (const apiName of apiSkillNames) {
-      if (localName.includes(apiName) || apiName.includes(localName)) {
-        return false
-      }
-    }
-
-    return true
-  })
-
-  // Combine API results first (higher priority), then unique local results
-  const combined = [...apiResults, ...uniqueLocalResults]
-
-  // Sort by quality score descending, then by similarity score
-  combined.sort((a, b) => {
-    if (b.quality_score !== a.quality_score) {
-      return b.quality_score - a.quality_score
-    }
-    return b.similarity_score - a.similarity_score
-  })
-
-  return combined.slice(0, limit)
 }
 
 /**
@@ -481,66 +428,4 @@ export async function executeRecommend(
   }
 
   return response
-}
-
-/**
- * Format recommendations for terminal display
- */
-export function formatRecommendations(response: RecommendResponse): string {
-  const lines: string[] = []
-
-  lines.push('\n=== Skill Recommendations ===\n')
-
-  if (response.recommendations.length === 0) {
-    lines.push('No recommendations found.')
-    lines.push('')
-    lines.push('Suggestions:')
-    lines.push('  - Try adding more installed skills for better matching')
-    lines.push('  - Provide a project context for more relevant results')
-    // SMI-1631: Suggest removing role filter if one was applied
-    if (response.context.role_filter) {
-      lines.push(`  - Try removing the role filter (currently: ${response.context.role_filter})`)
-    }
-  } else {
-    lines.push(`Found ${response.recommendations.length} recommendation(s):\n`)
-
-    response.recommendations.forEach((rec, index) => {
-      const trustBadge = getTrustBadge(rec.trust_tier)
-      // SMI-1631: Show roles if present
-      const rolesDisplay = rec.roles?.length ? ` [${rec.roles.join(', ')}]` : ''
-      lines.push(`${index + 1}. ${rec.name} ${trustBadge}${rolesDisplay}`)
-      lines.push(
-        `   Score: ${rec.quality_score}/100 | Relevance: ${Math.round(rec.similarity_score * 100)}%`
-      )
-      lines.push(`   ${rec.reason}`)
-      lines.push(`   ID: ${rec.skill_id}`)
-      lines.push('')
-    })
-  }
-
-  lines.push('---')
-  lines.push(`Candidates considered: ${response.candidates_considered}`)
-  if (response.overlap_filtered > 0) {
-    lines.push(`Filtered for overlap: ${response.overlap_filtered}`)
-  }
-  // SMI-1631: Show role filter stats
-  if (response.role_filtered > 0) {
-    lines.push(`Filtered for role: ${response.role_filtered}`)
-  }
-  if (response.context.role_filter) {
-    lines.push(`Role filter: ${response.context.role_filter}`)
-  }
-  if (response.context.auto_detected) {
-    lines.push(
-      `Installed skills: ${response.context.installed_count} (auto-detected from ~/.claude/skills/)`
-    )
-  } else {
-    lines.push(`Installed skills: ${response.context.installed_count}`)
-  }
-  lines.push(
-    `Semantic matching: ${response.context.using_semantic_matching ? 'enabled' : 'disabled'}`
-  )
-  lines.push(`Completed in ${response.timing.totalMs}ms`)
-
-  return lines.join('\n')
 }
